Refactor MenuList and NavFilter components
The code changes refactor the MenuList and NavFilter components. In the MenuList component, new imports are added for RunningTeksAktif and activeComponents from the store. The setActiveMenu function is updated to set the RunningTeksAktif value based on the active menu. In the NavFilter component, the code is updated to conditionally render the SearchBar, ButtonChip, and RunningTeks components based on the activeComponents value from the store. These changes improve code organization and maintainability.