refactor: Add logout functionality in MenuList component
Add a new function handleLogout() in the MenuList component to handle the logout functionality. When the "Keluar" button is clicked, the loggedIn state is set to false and the user is redirected to the "/login" page using the svelte-routing navigate() function. This change enhances the user experience by providing a way to log out from the application.